# config/supla-api-docs.yaml
openapi: 3.0.0
servers:
  - url: 'https://cloud.supla.org/api/v2.3.0'
info:
  version: "2.3.0"
  title: SUPLA Cloud API
paths:
...
  /channels:
    get:
      tags:
        - Channels
      summary: Get channels list
      operationId: getChannels
      parameters:
        - name: include
          in: query
          description: Specify what extra fields to include in the response.
          style: form
          schema:
            type: array
            items:
              type: string
              enum:
                - iodevice
                - location
                - connected
                - state
        - name: function
          in: query
          style: form
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ChannelFunctionEnumNames'
        - name: io
          in: query
          description: Return only `input` or `output` channels.
          schema:
            type: string
            enum:
              - input
              - output
        - name: hasFunction
          in: query
          description: >-
            Return only channels with (`true`) or without (`false`) chosen
            functions.
          schema:
            type: boolean
      responses:
        '200':
          description: List of channels owned by the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
...
  /channel-groups:
    get:
      tags:
        - Channel Groups
      summary: Get channels list
      operationId: getChannelGroups
      parameters:
        - name: include
          in: query
          description: Specify what extra fields to include in the response.
          style: form
          schema:
            type: array
            items:
              type: string
              enum:
                - channels
      responses:
        '200':
          description: List of channels oroups owned by the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelGroup'
    post:
      summary: Create a new channel group
      operationId: createChannelGroup
      tags:
        - Channel Groups
      responses:
        '201':
          description: Newely created channel group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelGroup'
      requestBody:
        $ref: '#/components/requestBodies/ChannelGroupRequest'
...
  /iodevices:
    get:
      tags:
        - IO Devices
      summary: Get IO Devices
      operationId: getIoDevices
      parameters:
        - name: include
          in: query
          description: Specify what extra fields to include in the response.
          style: form
          schema:
            type: array
            items:
              type: string
              enum:
                - channels
                - location
                - originalLocation
                - connected
                - schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
...
  /locations:
    get:
      tags:
        - Locations
      summary: Get locations list
      operationId: getLocations
      parameters:
        - name: include
          in: query
          description: Specify what extra fields to include in the response.
          required: false
          style: form
          schema:
            type: array
            items:
              type: string
              enum:
                - channels
                - iodevices
                - accessids
                - channelGroups
                - password
      responses:
        '200':
          description: List of locations owned by the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
    post:
      summary: Create a new location
      operationId: createLocation
      tags:
        - Locations
      responses:
        '201':
          description: Newely created location with automatic name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
  '/locations/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: include
        in: query
        description: Specify what extra fields to include in the response.
        required: false
        style: form
        schema:
          type: array
          items:
            type: string
            enum:
              - channels
              - iodevices
              - accessids
              - channelGroups
              - password
    get:
      summary: Get location by ID
      tags:
        - Locations
      operationId: getLocation
      responses:
        '200':
          description: Get location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
    put:
      summary: Update location
      tags:
        - Locations
      operationId: updateLocation
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
      requestBody:
        content:
          application/json:
            schema:
              title: Location update request
              properties:
                enabled:
                  type: boolean
                caption:
                  type: string
                password:
                  type: string
                  description: Provide new password if you want to change it.
                accessIdsIds:
                  type: array
                  items:
                    type: integer
                  description: Access Identifiers identifiers to assign to this location.
              type: object
        required: true
    delete:
      summary: Delete location
      tags:
        - Locations
      operationId: deleteLocation
      responses:
        '204':
          description: Location has been successfully deleted
...
security:
  - BearerAuth: []
  - OAuth2: []
components:
  schemas:
    AccessIdentifier:
      properties:
        id:
          type: integer
          description: Access Identifier identifier
        caption:
          type: string
          description: Location caption
        enabled:
          type: boolean
          description: '`true` if the location is enabled, `false` otherwise'
        locationsIds:
          type: array
          items:
            type: integer
          description: array containing the location idenfifiers assigned to this access ID
        clientAppsIds:
          type: array
          items:
            type: integer
          description: >-
            array containing the client apps idenfifiers assigned to this access
            ID
        locations:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/Location'
        clientApps:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/ClientApp'
      type: object
...
    Channel:
      properties:
        id:
          type: integer
          description: Channel identifier
        channelNumber:
          type: integer
          description: Channel ordinal number in its IO Device
        caption:
          type: string
          description: Channel caption
        type:
          $ref: '#/components/schemas/ChannelType'
        function:
          $ref: '#/components/schemas/ChannelFunction'
        param1:
          $ref: '#/components/schemas/ChannelParam'
        param2:
          $ref: '#/components/schemas/ChannelParam'
        param3:
          $ref: '#/components/schemas/ChannelParam'
        textParam1:
          type: string
        textParam2:
          type: string
        textParam3:
          type: string
        altIcon:
          type: integer
          description: >-
            Chosen alternative icon idenifier. Should not be greater than
            `funciton.maxAlternativeIconIndex`
        hidden:
          type: boolean
          description: Whether this channel is shown on client apps or not
        inheritedLocation:
          type: boolean
          description: >-
            Whether this channel inherits its IO Device's location (`true`) or
            not (`false`)
        iodeviceId:
          type: integer
        locationId:
          type: integer
        functionId:
          type: integer
          example: 60
        typeId:
          type: integer
          example: 1000
        userIconId:
          type: integer
          example: 100
        iodevice:
          $ref: '#/components/schemas/Device'
        location:
          $ref: '#/components/schemas/Location'
        connected:
          type: boolean
        state:
          $ref: '#/components/schemas/ChannelState'
        supportedFunctions:
          type: array
          items:
            $ref: '#/components/schemas/ChannelFunction'
      type: object
    ChannelState:
      description: >-
        The channel's state. The value and the format depends on the channel's function.
        Read more on [Github Wiki](https://github.com/SUPLA/supla-cloud/wiki/Channel-Functions-states).
      oneOf:
        - $ref: '#/components/schemas/BrightnessState'
        - $ref: '#/components/schemas/ColorAndBrightnessState'
        - $ref: '#/components/schemas/ColorState'
        - $ref: '#/components/schemas/DepthState'
        - $ref: '#/components/schemas/DistanceState'
        - $ref: '#/components/schemas/DoubleState'
        - $ref: '#/components/schemas/ElectricityMeterState'
        - $ref: '#/components/schemas/HumidityState'
        - $ref: '#/components/schemas/ImpulseCounterState'
        - $ref: '#/components/schemas/OnState'
        - $ref: '#/components/schemas/HiState'
        - $ref: '#/components/schemas/PartialOpenState'
        - $ref: '#/components/schemas/RollerShutterState'
        - $ref: '#/components/schemas/TemperatureAndHumidityState'
        - $ref: '#/components/schemas/TemperatureState'
...
    ChannelGroup:
      properties:
        id:
          type: integer
          description: Channel group identifier
          example: 1
        hidden:
          type: boolean
          description: Whether this channel group is shown on client apps or not
        caption:
          type: string
          description: Channel caption
        function:
          $ref: '#/components/schemas/ChannelFunction'
        functionId:
          type: integer
          example: 60
        userIconId:
          type: integer
        locationId:
          type: integer
        location:
          $ref: '#/components/schemas/Location'
        channelIds:
          type: array
          items:
            type: integer
        channels:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/Channel'
        altIcon:
          type: integer
          description: >-
            Chosen alternative icon idenifier. Should not be greater than
            `funciton.maxAlternativeIconIndex`
      type: object
    ChannelGroupRequest:
      required:
        - channelIds
      properties:
        caption:
          type: string
        altIcon:
          type: integer
        userIconId:
          type: integer
        hidden:
          type: boolean
        locationId:
          type: integer
        channelIds:
          type: array
          items:
            type: integer
      type: object
    ChannelType:
      properties:
        id:
          type: integer
          enum:
            - 1000
            - 1010
            - 1020
            - 1500
            - 2000
            - 2010
            - 2020
            - 2900
            - 3000
            - 3010
            - 3022
            - 3020
            - 3032
            - 3030
            - 3034
            - 3036
            - 3038
            - 3042
            - 3044
            - 3048
            - 3050
            - 3100
            - 4000
            - 4010
            - 4020
          description: Channel type identifier
        name:
          type: string
          enum:
            - SENSORNO
            - SENSORNC
            - DISTANCESENSOR
            - CALLBUTTON
            - RELAYHFD4
            - RELAYG5LA1A
            - RELAY2XG5LA1A
            - RELAY
            - THERMOMETERDS18B20
            - DHT11
            - DHT21
            - DHT22
            - AM2301
            - AM2302
            - THERMOMETER
            - HUMIDITYSENSOR
            - HUMIDITYANDTEMPSENSOR
            - WINDSENSOR
            - PRESSURESENSOR
            - RAINSENSOR
            - WEIGHTSENSOR
            - WEATHER_STATION
            - DIMMER
            - RGBLEDCONTROLLER
            - DIMMERANDRGBLED
        caption:
          type: string
          example: Sensor (normal open)
        output:
          type: boolean
          example: false
          description: >-
            Whether the channel is output type (i.e. can take action) or input
            (i.e. provide data)
      type: object
    ChannelFunction:
      properties:
        id:
          type: integer
          example: 60
          enum:
            - 0
            - 10
            - 20
            - 30
            - 40
            - 42
            - 45
            - 50
            - 60
            - 70
            - 80
            - 90
            - 100
            - 110
            - 120
            - 130
            - 140
            - 180
            - 190
            - 200
            - 210
            - 220
            - 230
            - 240
            - 250
            - 260
            - 270
            - 280
            - 290
            - 300
        name:
          $ref: '#/components/schemas/ChannelFunctionEnumNames'
        caption:
          type: string
          example: Gate opening sensor
        maxAlternativeIconIndex:
          type: integer
        possibleActions:
          type: array
          description: What can you with with this channel?
          items:
            $ref: '#/components/schemas/ChannelFunctionAction'
        possibleVisualStates:
          type: array
          description: Possible visual states of channel with this function. Ordered.
          items:
            type: string
            enum:
              - opened
              - closed
              - partially_closed
              - default
              - empty
              - full
              - revealed
              - shut
              - on
              - off
              - rgb_on_dim_on
              - rgb_on_dim_off
              - rgb_off_dim_off
              - rgb_off_dim_on
      type: object
    ChannelFunctionEnumNames:
      type: string
      example: OPENINGSENSOR_GATE
      enum:
        - NONE
        - CONTROLLINGTHEGATEWAYLOCK
        - CONTROLLINGTHEGATE
        - CONTROLLINGTHEGARAGEDOOR
        - THERMOMETER
        - HUMIDITY
        - HUMIDITYANDTEMPERATURE
        - OPENINGSENSOR_GATEWAY
        - OPENINGSENSOR_GATE
        - OPENINGSENSOR_GARAGEDOOR
        - NOLIQUIDSENSOR
        - CONTROLLINGTHEDOORLOCK
        - OPENINGSENSOR_DOOR
        - CONTROLLINGTHEROLLERSHUTTER
        - OPENINGSENSOR_ROLLERSHUTTER
        - POWERSWITCH
        - LIGHTSWITCH
        - DIMMER
        - RGBLIGHTING
        - DIMMERANDRGBLIGHTING
        - DEPTHSENSOR
        - DISTANCESENSOR
        - OPENINGSENSOR_WINDOW
        - MAILSENSOR
        - WINDSENSOR
        - PRESSURESENSOR
        - RAINSENSOR
        - WEIGHTSENSOR
        - WEATHER_STATION
        - STAIRCASETIMER
...
    ChannelParam:
      type: integer
      description: >-
        Configuration of the channel behavior. Value meaning depends on the
        chosen channel function. See more details on [Github
        Wiki](https://github.com/SUPLA/supla-cloud/wiki/Channel-Functions-parameters).
    Device:
      properties:
        id:
          type: integer
          description: Device identifier
        name:
          type: string
        enabled:
          type: boolean
        comment:
          type: string
        regDate:
          type: string
          format: date-time
        regIpv4:
          type: integer
          format: int64
        lastConnected:
          type: string
          format: date-time
        lastIpv4:
          type: integer
          format: int64
        softwareVersion:
          type: string
        gUIDString:
          type: string
        locationId:
          type: integer
        originalLocationId:
          type: integer
        channelsIds:
          type: array
          items:
            type: integer
        connected:
          type: boolean
        location:
          $ref: '#/components/schemas/Location'
        originalLocation:
          $ref: '#/components/schemas/Location'
        channels:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/Channel'
        schedules:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/Schedule'
      type: object
...
    Location:
      properties:
        id:
          type: integer
          description: Location identifier
        caption:
          type: string
          description: Location caption
        enabled:
          type: boolean
          description: '`true` if the location is enabled, `false` otherwise'
        iodevicesIds:
          type: array
          items:
            type: integer
          description: >-
            array containing the IO Devices identifiers assigned to this
            location
        channelGroupsIds:
          type: array
          items:
            type: integer
          description: >-
            array containing the Channel groups identifiers assigned to this
            location
        channelsIds:
          type: array
          items:
            type: integer
          description: >-
            array containing the Channels identifiers expicitely assigned to
            this location
        accessIdsIds:
          type: array
          items:
            type: integer
          description: >-
            array containing the Access Identifiers identifiers assigned to this
            location
        password:
          type: string
          description: >-
            Location password (plain text). Returned only if requested by the
            `include` parameter.
        iodevices:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/Device'
        channelGroups:
          type: array
          description: Returned only if requested by the `include` parameter.
          example: []
          items:
            $ref: '#/components/schemas/ChannelGroup'
        accessIds:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/AccessIdentifier'
        channels:
          type: array
          description: Returned only if requested by the `include` parameter.
          items:
            $ref: '#/components/schemas/Channel'
      type: object
    Schedule:
      properties:
        id:
          type: integer
          description: Schedule identifier
        timeExpression:
          type: string
          description: >-
            Schedule time expression in crontab notation (with some custom
            additions).
          example: '*/5 * * * *'
        action:
          $ref: '#/components/schemas/ChannelFunctionAction'
        actionParam:
          example: {}
          description: Depends on the action.
        mode:
          type: string
          enum:
            - minutely
            - hourly
            - daily
            - once
        dateStart:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        enabled:
          type: boolean
        caption:
          type: string
        retry:
          type: boolean
        channelId:
          type: integer
        actionId:
          type: integer
        channel:
          $ref: '#/components/schemas/Channel'
        closestExecutions:
          type: array
          description: >-
            Returned only if requested by the `include` parameter. Contains two
            arrays of maximum 3 closest past and future executions.
          items:
            properties:
              past:
                type: array
                items:
                  properties:
                    plannedTimestamp:
                      type: string
                      format: date-time
                    resultTimestamp:
                      type: string
                      format: date-time
                    failed:
                      type: boolean
                      example: false
                    result:
                      properties:
                        caption:
                          type: string
                          example: Successful
                      type: object
                  type: object
              future:
                type: array
                items:
                  properties:
                    plannedTimestamp:
                      type: string
                      format: date-time
                  type: object
            type: object
      type: object
...
  requestBodies:
    ChannelGroupRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChannelGroupRequest'
      required: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'https://cloud.supla.org/oauth/v2/auth'
          tokenUrl: 'https://cloud.supla.org/oauth/v2/token'
          scopes:
            accessids_r: Access Identifiers (Read)
            accessids_rw: Access Identifiers (Read/Write)
            account_r: User Account (Read)
            account_rw: User Account (Read/Write)
            channels_r: Channels (Read)
            channels_rw: Channels (Read/Write)
            channels_ea: Channels (Execute Actions)
            channels_files: Channels (download files, e.g. icons, measurement logs)
            channelgroups_r: Channel Groups (Read)
            channelgroups_rw: Channel Groups (Read/Write)
            channelgroups_ea: Channel Groups (Execute Actions)
            clientapps_r: Client Apps (Read)
            clientapps_rw: Client Apps (Read/Write)
            directlinks_r: Direct Links (Read)
            directlinks_rw: Direct Links (Read/Write)
            iodevices_r: IO Devices (Read)
            iodevices_rw: IO Devices (Read/Write)
            locations_r: Locations (Read)
            locations_rw: Locations (Read/Write)
            schedules_r: Schedules (Read)
            schedules_rw: Schedules (Read/Write)
            offline_access: Request refresh token